Normal distribution implementation based on colt
Extends the Colt normal class with the distribution protocol. Adds a
convenience function to create normal distributions. Deprecates (using a
macro added to incanter.internal) the various *-normal functions, with
the exception of quartile-normal, as I do net have an alternative. A
more complete deprecation of those functions would implement them in
terms of the distribution protocol, but I'll leave that for another day.