Added pdf and cdf of other distributions from incanter.stats
Bootstrapped from the colt implementation for some. For others, namely F,
began using defrecord. draw() and support() do not work for all dists,
as the latter's return format is still to be decided, as is whether to
have native implementations, and there is an issue with :lower-tail
given the current, intended API.