incanter.io/read-dataset reads on a per-line basis
read-dataset used CSVReader.readAll which parsed the whole file at once.
In case of larger CSV files it caused some major headaches for the GC. This
patch makes CSVReader work on a per-line basis by using CSVReader.read and
fully parsing each line before moving to a next one.

A following benchmark was used to measure effects of this patch. 1k.csv is
a 1k lines long dataset with 700 numerical features. It's around 1.7MB.

    cat <<EOF | JAVA_OPTS=-Xmx100M incanter
    (do
      (use '(incanter io core))
      (time
        (doall (map
                 (fn [_]
                     (nrow (incanter.io/read-dataset "1k.csv")))
                 (range 5)))))
    EOF

Original read-dataset needed 13.5 seconds. Profiling showed that a
significant part of this time was spent in GC. This patch reduces
execution time to 7.3s by significantly reducing the number of String
and char[] instances stored simultaneously on the heap. Memory consumption
is reduced as well.

It's also worth mentioning that when 1k.csv aws replaced with a 3 times
larger 3k.csv the original read-dataset ran out of memory. This patch
solves the problem.

Signed-off-by: Jan Stępień <jstepien@users.sourceforge.net>