Correct problems found via Eastwood Clojure lint tool in incanter-core
Replace doall with dorun in a place where the return value is
discarded.  doall holds onto the head of the result sequence
unnecessarily in this case.

One arity of fn auto-correlation should have been calling its larger
arity version, but was not.

Replace deprecated clojure.core/replicate with repeat

Correct location of doc string for several functions.

Several tests had misplaced parentheses, were missing (is ...)  around
(= ...), or missing (= ...) inside of (is ...)

Changed test that expected return value of (safe-div 0) to be NaN so
that it expected Double/POSITIVE_INFINITY instead.  That seems
consistent with other tests.

(= NaN NaN) is always false in Java and Clojure.  Changed tests that
expected a NaN result to use Java Double's isNaN method.

Two deftest's in stats_test.clj had same name covariance-test, causing
first one to be replaced and its tests never run.  Renamed second one
to covariance-test-2 so all tests are run.

There is one test that now fails after adding (is ...) around (= ...).
I did not try to correct it, as I was not sure what the expected
return value should be.