Sel returns dataset when :rows :all specified
Before the fix a simple list was returned for datasets with a single row
when :cols :all :rows :all was used. Although sel is supposed to return
simple lists only when at least one of the query params :rows and :cols
is an atomic object (index or name).

Fixes github issue #235