Adds header function option to read-dataset.
If file has a header line, you can pass in a function that is applied to
all items in the header. The header function is set with the :header-fn option.
The header function is only called if :header true is also passed as an
option. The default header function is keyword, thus maintaining backward
compatibility.
If :header is false, the header function is ignored.

The following are the same:
(read-dataset filename :header true)
(read-dataset :header true :header-fn keyword)

This option allows flexibility for when your headers need extra
manipulation.
For example, if you want your headers lowercase, you can do the following:
(read-dataset :header true :header-fn (comp keyword clojure.string/lower-case))