Refactor io/read-dataset.
Function read-dataset now accepts two new function parameters:
    row-fn - is applied to each row before constructing the dataset.
    dataset-fn - is applied to entire sequence of rows.

Only a few previous option parameters were retained.
   delim
   quote
   header
And a new option skip-blanklines was added, which behaves the same as
   the original function. All option default values tried to conform to
   previous behavior or to the the R function read.table.

All previous behavior found in the function can be obtained by using the
rwo-fn and dataset-fn options. Examples on how to obtain the previous
behavior can be found in the unit tests.