Change get-dataset to use new read-dataset API.
Adds tests for incanter.datasets namespace. Includes test for all
Incanter provided datasets.
Only the iran-election dataset is not tested due to unmatched column
counts. This dataset is unreadable by the CSV parser and must be fixed.

The incanter.datasets tests use the local version of Incanter datasets
instead of retrieving the datasets from the repo. This makes running the
tests faster and facilitates possible format changes to the datasets and
testing the changes.

Formatting of many datasets was fixed so they could be read by
clojure.data.csv/read-csv. Mainly extra spaces were removed in files
with \space (single space) as the delimiter.