Fix errors in charts unit tests.
Tests now read from the local datasets instead of retrieving them from
the repo.

Columns that are numbers are explicitly converted to double.
The bar-chart and line-chart functions expect values as numbers (more
specifically the underlying jfree DefaultCategoryDataset being
constructed), this caused errors since all fields were strings.

The exact issue was that the `addValue` method of DefaultCategoryDataset(
http://www.jfree.org/jfreechart/api/javadoc/org/jfree/data/category/DefaultCategoryDataset.html#addValue-double-java.lang.Comparable-java.lang.Comparable-)
expects either a Number or double as the value, therefore we have to
convert the fields accordingly.

The ideal way would be to pass in a :row-fn when reading the dataset to
convert columns to their corresponding types.