Exclude `clojure.core/update` in `incanter.core`
Clojure 1.7 introduces `clojure.core/update`, which means that prior
to this commit `incanter.core/update` would shadow the `clojure.core`
var and throw a warning.  This commit uses:

    (:refer-clojure :exclude [...])

...to make the `clojure.core` var not in scope in `incanter.core`,
preventing the warning.  `incanter.core` never refers the
`clojure.core` var, so there should not be any negative effects of
this exclusion.