Fixed heat-map artifacts, expanded API
For certain ranges (empirically determined
to be  xmax - xmin > 100), the plotting
for heat-maps became sparse.  Compared
to a high-resolution java demo, this
seemed counterintuitive.  Visual artifacts
appeared due to the background showing
through spaces in between the "blocks"
serving as pixels from the renderer.

incanter.core/grid-apply was part of
the culprit, in addition to not
dynamically resizing the block-size
for the XYBlockRenderer.

We now have an additional function
arity for grid-apply, which allows
callers to specify x-res and y-res,
respective sample widths for each
axis.  Alone, with a good guess
at a high resolution sample, this
fixes the problem.

However, we should be able to
provide "blocky" heatmaps with
samples if we can scale up the
block size.  incanter.charts/heat-map
now accepts :x-res and :y-res
as optional arguments, in addition
to :auto-scale?.  Auto scaling
- true by default - will scale
up the block size to fit the
available samples to provide
a continous surface.  Callers
can disable it if the intent is
to provide a "sparse" surface
sampled at discrete points.